+++ autonomy/deploy/generators/docker_compose/base.py
@@ -21,7 +21,7 @@
 from pathlib import Path
 from typing import Dict, Optional
 
-from aea.configurations.constants import DEFAULT_PRIVATE_KEY_FILE
+from aea.configurations.constants import DEFAULT_LEDGER, PRIVATE_KEY_PATH_SCHEMA
 from docker import from_env
 
 from autonomy.constants import (
@@ -41,6 +41,7 @@ from autonomy.deploy.constants import (
     DEPLOYMENT_KEY_DIRECTORY,
     INFO,
     KEY_SCHEMA_PRIVATE_KEY,
+    KEY_SCHEMA_TYPE,
 )
 from autonomy.deploy.generators.docker_compose.templates import (
     ABCI_NODE_TEMPLATE,
@@ -67,6 +68,7 @@ def build_tendermint_node_config(
     node_id: int,
     dev_mode: bool = False,
     log_level: str = INFO,
+    tendermint_ports: Optional[Dict[int, int]] = None,
 ) -> str:
     """Build tendermint node config for docker compose."""
 
@@ -82,6 +84,15 @@ def build_tendermint_node_config(
     if dev_mode:
         config += "      - ./persistent_data/tm_state:/tm_state:Z"
         config = config.replace("DEV_MODE=0", "DEV_MODE=1")
+
+    if tendermint_ports is not None:
+        port_mappings = map(
+            lambda x: PORT_MAPPING_CONFIG.format(host_port=x[0], container_port=x[1]),
+            tendermint_ports.items(),
+        )
+        port_config = "\n".join([PORTS, *port_mappings])
+        config += port_config
+        config += "\n"
 
     return config
 
@@ -123,6 +134,7 @@ def build_agent_config(  # pylint: disable=too-many-arguments
         )
         port_config = "\n".join([PORTS, *port_mappings])
         config += port_config
+        config += "\n"
 
     return config
 
@@ -207,6 +219,13 @@ class DockerComposeGenerator(BaseDeploymentGenerator):
                     node_id=i,
                     dev_mode=self.dev_mode,
                     log_level=self.service_builder.log_level,
+                    tendermint_ports=(
+                        self.service_builder.service.deployment_config.get(
+                            "tendermint", {}
+                        )
+                        .get("ports", {})
+                        .get(i)
+                    ),
                 )
                 for i in range(self.service_builder.service.number_of_agents)
             ]
@@ -242,9 +261,9 @@ class DockerComposeGenerator(BaseDeploymentGenerator):
         keys_dir = self.build_dir / DEPLOYMENT_KEY_DIRECTORY
         for x in range(self.service_builder.service.number_of_agents):
             path = keys_dir / DEPLOYMENT_AGENT_KEY_DIRECTORY_SCHEMA.format(agent_n=x)
-            keys_file = path / DEFAULT_PRIVATE_KEY_FILE
+            ledger = self.service_builder.keys[x].get(KEY_SCHEMA_TYPE, DEFAULT_LEDGER)
             key = self.service_builder.keys[x][KEY_SCHEMA_PRIVATE_KEY]
-
+            keys_file = path / PRIVATE_KEY_PATH_SCHEMA.format(ledger)
             path.mkdir()
             with keys_file.open(mode="w", encoding=DEFAULT_ENCODING) as f:
                 f.write(key)