+++ packages/valory/skills/abstract_round_abci/tests/test_behaviours_utils.py
@@ -2205,6 +2205,12 @@ class TestBaseBehaviour:
                 self.behaviour.context.params.tendermint_com_url + "/hard_reset",
                 parameters=expected_parameters,
             )
+
+            should_be_healthy = isinstance(reset_response, dict) and reset_response.get(
+                "status", False
+            )
+            assert self.behaviour._is_healthy is should_be_healthy
+
             # perform the last iteration which also returns the result
             try:
                 next(reset)
@@ -2226,6 +2232,7 @@ class TestBaseBehaviour:
                         tm_recovery_params.reset_from_round
                         == self.behaviour.matching_round.auto_round_id()
                     )
+                    assert not self.behaviour._is_healthy
             else:
                 pytest.fail("`reset_tendermint_with_wait` did not finish!")
 
@@ -2324,8 +2331,16 @@ class TestTmManager:
         ):
             self.tm_manager.act_wrapper()
 
+    @given(latest_block_height=st.integers(min_value=0))
     @pytest.mark.parametrize(
         "acn_communication_success",
+        (
+            True,
+            False,
+        ),
+    )
+    @pytest.mark.parametrize(
+        "gentle_reset_attempted",
         (
             True,
             False,
@@ -2342,16 +2357,39 @@ class TestTmManager:
     )
     def test_handle_unhealthy_tm(
         self,
+        latest_block_height: int,
         acn_communication_success: bool,
+        gentle_reset_attempted: bool,
         tm_reset_success: bool,
         num_active_peers: Optional[int],
     ) -> None:
         """Test _handle_unhealthy_tm."""
 
+        self.tm_manager.gentle_reset_attempted = gentle_reset_attempted
+        self.tm_manager.context.state.round_sequence.height = latest_block_height
+
         def mock_sleep(_seconds: int) -> Generator:
             """A method that mocks sleep."""
             return
             yield
+
+        def dummy_do_request(*_: Any) -> Generator[None, None, MagicMock]:
+            """Dummy `_do_request` method."""
+            yield
+            return mock.MagicMock()
+
+        def dummy_get_status(*_: Any) -> Generator[None, None, MagicMock]:
+            """Dummy `_get_status` method."""
+            yield
+            return mock.MagicMock(
+                body=json.dumps(
+                    {
+                        "result": {
+                            "sync_info": {"latest_block_height": latest_block_height}
+                        }
+                    }
+                ).encode()
+            )
 
         gen = self.tm_manager._handle_unhealthy_tm()
         with mock.patch.object(
@@ -2364,15 +2402,24 @@ class TestTmManager:
             side_effect=yield_and_return_int_wrapper(num_active_peers),
         ), mock.patch.object(
             self.tm_manager, "sleep", side_effect=mock_sleep
-        ), mock.patch(
-            "sys.exit"
-        ) as mock_sys_exit, mock.patch.object(
+        ), mock.patch.object(
             BaseBehaviour,
             "request_recovery_params",
             side_effect=dummy_generator_wrapper(acn_communication_success),
+        ), mock.patch.object(
+            BaseBehaviour, "_do_request", new_callable=lambda *_: dummy_do_request
+        ), mock.patch.object(
+            BaseBehaviour, "_get_status", new_callable=lambda *_: dummy_get_status
         ):
             next(gen)
-            next(gen)
+
+            if not gentle_reset_attempted:
+                next(gen)
+                assert self.tm_manager.gentle_reset_attempted
+                with pytest.raises(StopIteration):
+                    next(gen)
+                assert not self.tm_manager.gentle_reset_attempted
+                return
 
             if not acn_communication_success:
                 with pytest.raises(StopIteration):
@@ -2382,12 +2429,6 @@ class TestTmManager:
             next(gen)
             with pytest.raises(StopIteration):
                 next(gen)
-
-            if (
-                num_active_peers is not None
-                and num_active_peers < self._DUMMY_CONSENSUS_THRESHOLD
-            ):
-                mock_sys_exit.assert_called()
 
     @pytest.mark.parametrize(
         "expected_reset_params",