+++ tests/test_autonomy/test_cli/test_deploy/test_from_token.py
@@ -29,7 +29,6 @@ from aea.cli.registry.settings import REMOTE_IPFS
 from aea_test_autonomy.fixture_helpers import registries_scope_class  # noqa: F401
 
 from autonomy.chain.exceptions import FailedToRetrieveComponentMetadata
-from autonomy.cli.helpers.deployment import BadFunctionCallOutput
 
 from tests.conftest import ROOT_DIR, skip_docker_tests
 from tests.test_autonomy.test_chain.base import BaseChainInteractionTest
@@ -117,7 +116,9 @@ class TestFromToken(BaseChainInteractionTest):
         with mock.patch(
             "autonomy.cli.helpers.deployment.fetch_service_ipfs",
             return_value=service_dir,
-        ), run_deployment_patch, build_image_patch, default_remote_registry_patch, default_ipfs_node_patch, ipfs_resolve_patch:
+        ), (run_deployment_patch), (
+            build_image_patch
+        ), default_remote_registry_patch, default_ipfs_node_patch, ipfs_resolve_patch:
             result = self.run_cli(
                 (
                     str(self.token),
@@ -152,7 +153,9 @@ class TestFromToken(BaseChainInteractionTest):
         with mock.patch(
             "autonomy.cli.helpers.deployment.fetch_service_ipfs",
             return_value=service_dir,
-        ), run_deployment_patch as rdp, build_image_patch, default_remote_registry_patch, default_ipfs_node_patch, ipfs_resolve_patch:
+        ), run_deployment_patch as rdp, (
+            build_image_patch
+        ), default_remote_registry_patch, default_ipfs_node_patch, ipfs_resolve_patch:
             result = self.run_cli(
                 (str(self.token), str(self.keys_file), "--kubernetes")
             )
@@ -171,7 +174,9 @@ class TestFromToken(BaseChainInteractionTest):
     def test_fail_on_chain_resolve_connection_error(self) -> None:
         """Run test."""
 
-        with run_deployment_patch, build_image_patch, default_remote_registry_patch, default_ipfs_node_patch, ipfs_resolve_patch, mock.patch(
+        with (run_deployment_patch), (build_image_patch), (
+            default_remote_registry_patch
+        ), default_ipfs_node_patch, ipfs_resolve_patch, mock.patch(
             "autonomy.cli.helpers.deployment.resolve_component_id",
             side_effect=FailedToRetrieveComponentMetadata(
                 "Error connecting RPC endpoint"
@@ -190,9 +195,11 @@ class TestFromToken(BaseChainInteractionTest):
     def test_fail_on_chain_resolve_bad_contract_call(self) -> None:
         """Run test."""
 
-        with run_deployment_patch, build_image_patch, default_remote_registry_patch, default_ipfs_node_patch, ipfs_resolve_patch, mock.patch(
+        with (run_deployment_patch), (build_image_patch), (
+            default_remote_registry_patch
+        ), default_ipfs_node_patch, ipfs_resolve_patch, mock.patch(
             "autonomy.cli.helpers.deployment.resolve_component_id",
-            side_effect=BadFunctionCallOutput,
+            side_effect=Exception,
         ):
             result = self.run_cli(
                 (