+++ tests/test_autonomy/test_images/test_runtime.py
@@ -38,7 +38,7 @@ from autonomy.constants import (
 )
 from autonomy.deploy.constants import DOCKERFILES
 
-from tests.conftest import ROOT_DIR, get_package_hash_from_latest_tag, skip_docker_tests
+from tests.conftest import ROOT_DIR, skip_docker_tests
 from tests.test_autonomy.test_images.base import BaseImageBuildTest
 
 
@@ -62,11 +62,10 @@ class TestOpenAutonomyBaseImage(BaseImageBuildTest):
         """Setup class."""
         super().setup_class()
 
-        cls.agent = str(
-            AGENT.with_hash(
-                get_package_hash_from_latest_tag(package=AGENT.to_uri_path)
-            ).public_id
-        )
+        # TODO: Replace this logic with logic to use hash from the previous release
+        # after v0.10.5.post1
+        package_hash = "bafybeifmmrwsjl7cnp4p4pjsn7qmkofmqoll5rl3cukwnlnlfcz64i553y"
+        cls.agent = str(AGENT.with_hash(package_hash).public_id)
 
     def test_image_fail(self) -> None:
         """Test image build fail."""
@@ -99,6 +98,7 @@ class TestOpenAutonomyBaseImage(BaseImageBuildTest):
         )
 
         assert success, output
+        assert "Successfully built the agent" in output
         assert f"Successfully tagged {self.tag}" in output
 
         # check runtime